Fix swiftlint warning
